Restructure vocabularies into separate classes
Fixes #333
Fixes #334
