+++ spec/cocina/generator/schema_spec.rb
@@ -67,7 +67,7 @@ RSpec.describe Cocina::Generator::Schema do
       Cocina::Models::AdminPolicy.new(
         externalIdentifier: 'druid:bc123df4567',
         label: 'My admin policy',
-        type: Cocina::Models::Vocab.admin_policy,
+        type: Cocina::Models::ObjectType.admin_policy,
         version: 1,
         administrative: {
           hasAdminPolicy: 'druid:bc123df4567',
@@ -90,7 +90,7 @@ RSpec.describe Cocina::Generator::Schema do
       Cocina::Models::AdminPolicy.new({
                                         externalIdentifier: 'druid:abc123',
                                         label: 'My admin policy',
-                                        type: Cocina::Models::Vocab.admin_policy,
+                                        type: Cocina::Models::ObjectType.admin_policy,
                                         version: 1,
                                         administrative: {}
                                       }, false, true)
