+++ spec/cocina/models/object_type_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe Cocina::Models::Vocab do
+RSpec.describe Cocina::Models::ObjectType do
   it 'provides vocab methods' do
     expect(described_class.admin_policy).to eq('http://cocina.sul.stanford.edu/models/admin_policy.jsonld')
     expect(described_class.webarchive_seed).to eq('http://cocina.sul.stanford.edu/models/webarchive-seed.jsonld')
